pyproject.toml
[tool.poetry]
name = "drf-practice"
version = "0.1.0"
description = "DRF Practice to learn"
authors = ["myname <your-email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
djangorestframework-camel-case = "^1.4.2"
drf-yasg = "^1.21.7"
django-filter = "^23.5"
django-model-utils = "^4.3.1"
gunicorn = "^21.2.0"
django-environ = "^0.11.2"
mysqlclient = "^2.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
